unit test for pbkdf2
